Fix expecttest accept on Python 3.8 and later
Fixes #59705.

Python 3.8 fixed tracebacks to report the beginning of the line
that raised an error, rather than the end.  This makes for a simpler
implementation (no more string reversing) but need to actually
implement.  This wasn't caught by tests because we hard coded line
numbers to do substitutions, so I also added a little smoketest to
detect future changes to traceback line number behavior.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 6265671168f401aafd5df21c56d1b931fd8fd3a2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/59709